refacot is null schema check
